backend/requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
requests
python-dotenv
SQLAlchemy
psycopg2-binary
python-multipart
openai
pathlib
pymupdf
redis
tiktoken